hoan-thanh-policies-category01

+++ app/Http/Controllers/GroupController.php
@@ -27,7 +27,7 @@ class GroupController extends Controller
      */
     public function index(Request $request)
     {
-        // $this->authorize('viewAny',Group::class);
+        $this->authorize('viewAny',Group::class);
         $groups = $this->groupService->paginate($request);
         return view('admin.groups.index',compact('groups'));
        
@@ -38,7 +38,7 @@ class GroupController extends Controller
      */
     public function create()
     {
-        // $this->authorize('create',Group::class);
+        $this->authorize('create',Group::class);
         return view('admin.groups.create');
     }
 
@@ -47,6 +47,7 @@ class GroupController extends Controller
      */
     public function store(StoreGroupRequest $request)
     {
+      
         try {
             $this->groupService->store($request);
             toast('Thêm Quyền Thành Công!', 'success', 'top-right');
@@ -73,7 +74,7 @@ class GroupController extends Controller
     public function edit(string $id)
     {
         
-        // $this->authorize('update',Group::class);
+        $this->authorize('update',Group::class);
         $group = $this->groupService->find($id);
         return view('admin.groups.edit', compact('group') );
     }
@@ -100,6 +101,7 @@ class GroupController extends Controller
      */
     public function destroy($id)
     {
+        $this->authorize('delete', Group::class);
         try {
             $this->groupService->forceDelete($id);
             toast('Nhóm Quyền Đã Được Xóa!', 'success', 'top-right');
